% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_equality.R
\name{check_equality}
\alias{check_equality}
\title{check_equality}
\usage{
check_equality(equality_table, pixels = seq(1, 2001, 10))
}
\arguments{
\item{equality_table}{}

\item{pixels}{}
}
\description{
check_equality
}
