% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_size_files.R
\name{equal_size_files}
\alias{equal_size_files}
\title{equal_size_files}
\usage{
equal_size_files(dir_A, dir_B, ...)
}
\arguments{
\item{dir_A}{}

\item{dir_B}{}

\item{...}{}
}
\description{
equal_size_files
}
