% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_doubles.R
\name{move_doubles_jpeg}
\alias{move_doubles_jpeg}
\title{move_doubles_jpeg}
\usage{
move_doubles_jpeg(equality_table, dir_doubles_backup, pixels = seq(1,
  2001, 10))
}
\arguments{
\item{equality_table}{test}

\item{dir_doubles_backup}{test}

\item{pixels}{test}
}
\description{
move_doubles_jpeg
}
