% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_empty_directories.R
\name{remove_empty_directories}
\alias{remove_empty_directories}
\title{remove_empty_directories}
\usage{
remove_empty_directories(x)
}
\arguments{
\item{x}{test}
}
\description{
remove_empty_directories
}
