% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_size_files.R
\name{equal_size_files}
\alias{equal_size_files}
\title{equal_size_files}
\usage{
equal_size_files(dir_with_doubles, dir_reference, ...)
}
\arguments{
\item{dir_with_doubles}{test}

\item{dir_reference}{test}

\item{...}{test}
}
\description{
equal_size_files
}
